Remove pull(ing) methods from container
- Move pullImage and dependent functions to a new file (`image.go`)
- Pulling image is handled by the service (using `ensureImageExists`)
  and thus there is no need for `pull` or `Pull` methods. It simplifies
  the code. This also mean that, when creating a container, if there is
  an error, even if it's an ErrNoImageFound, it won't try to pull it,
  the responsability is on the service.
