+++ docker/container.go
@@ -1,8 +1,6 @@
 package docker
 
 import (
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"io"
 	"math"
@@ -13,12 +11,9 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/docker/docker/pkg/promise"
 	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/docker/docker/pkg/term"
-	"github.com/docker/docker/reference"
-	"github.com/docker/docker/registry"
 	"github.com/docker/engine-api/client"
 	"github.com/docker/engine-api/types"
 	"github.com/docker/engine-api/types/container"
@@ -497,18 +492,8 @@ func (c *Container) createContainer(imageName, oldContainer string, configOverri
 
 	container, err := c.client.ContainerCreate(context.Background(), configWrapper.Config, configWrapper.HostConfig, configWrapper.NetworkingConfig, c.name)
 	if err != nil {
-		if client.IsErrImageNotFound(err) {
-			logrus.Debugf("Not Found, pulling image %s", configWrapper.Config.Image)
-			if err = c.pull(configWrapper.Config.Image); err != nil {
-				return nil, err
-			}
-			if container, err = c.client.ContainerCreate(context.Background(), configWrapper.Config, configWrapper.HostConfig, configWrapper.NetworkingConfig, c.name); err != nil {
-				return nil, err
-			}
-		} else {
-			logrus.Debugf("Failed to create container %s: %v", c.name, err)
-			return nil, err
-		}
+		logrus.Debugf("Failed to create container %s: %v", c.name, err)
+		return nil, err
 	}
 
 	return GetContainer(c.client, container.ID)
@@ -608,11 +593,6 @@ func (c *Container) Name() string {
 	return c.name
 }
 
-// Pull pulls the image the container is based on.
-func (c *Container) Pull() error {
-	return c.pull(c.service.serviceConfig.Image)
-}
-
 // Restart restarts the container if existing, does nothing otherwise.
 func (c *Container) Restart(timeout int) error {
 	container, err := c.findExisting()
@@ -659,65 +639,6 @@ func (c *Container) Log(follow bool) error {
 	logrus.WithFields(logrus.Fields{"Logger": l, "err": err}).Debug("c.client.Logs() returned error")
 
 	return err
-}
-
-func (c *Container) pull(image string) error {
-	return pullImage(c.client, c.service, image)
-}
-
-func pullImage(client client.APIClient, service *Service, image string) error {
-	distributionRef, err := reference.ParseNamed(image)
-	if err != nil {
-		return err
-	}
-
-	repoInfo, err := registry.ParseRepositoryInfo(distributionRef)
-	if err != nil {
-		return err
-	}
-
-	authConfig := service.context.AuthLookup.Lookup(repoInfo)
-
-	encodedAuth, err := encodeAuthToBase64(authConfig)
-	if err != nil {
-		return err
-	}
-
-	options := types.ImagePullOptions{
-		RegistryAuth: encodedAuth,
-	}
-	responseBody, err := client.ImagePull(context.Background(), distributionRef.String(), options)
-	if err != nil {
-		logrus.Errorf("Failed to pull image %s: %v", image, err)
-		return err
-	}
-	defer responseBody.Close()
-
-	var writeBuff io.Writer = os.Stdout
-
-	outFd, isTerminalOut := term.GetFdInfo(os.Stdout)
-
-	err = jsonmessage.DisplayJSONMessagesStream(responseBody, writeBuff, outFd, isTerminalOut, nil)
-	if err != nil {
-		if jerr, ok := err.(*jsonmessage.JSONError); ok {
-			// If no error code is set, default to 1
-			if jerr.Code == 0 {
-				jerr.Code = 1
-			}
-			fmt.Fprintf(os.Stderr, "%s", writeBuff)
-			return fmt.Errorf("Status: %s, Code: %d", jerr.Message, jerr.Code)
-		}
-	}
-	return err
-}
-
-// encodeAuthToBase64 serializes the auth configuration as JSON base64 payload
-func encodeAuthToBase64(authConfig types.AuthConfig) (string, error) {
-	buf, err := json.Marshal(authConfig)
-	if err != nil {
-		return "", err
-	}
-	return base64.URLEncoding.EncodeToString(buf), nil
 }
 
 func (c *Container) withContainer(action func(*types.ContainerJSON) error) error {
